# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SVM Scan API
  version: '1.0'
  description: |
    **SVM Scan** is a read-only REST API that serves indexed data for the
    Steem Virtual Machine chain — a Cosmos SDK chain with an integrated EVM.
    It's the backend for [svmscan.blazeapps.org](https://svmscan.blazeapps.org),
    but is itself a public, unauthenticated API that anyone can call directly.

    A background indexer walks the chain block-by-block into MongoDB
    (backfill from genesis, then a live tail of new blocks), and this API
    serves that indexed data. A few endpoints are deliberate exceptions that
    proxy the chain (or an external LCD/EVM RPC) live instead of reading
    from the index — each one says so explicitly in its description below.

    ## Conventions

    - **Pagination**: every list endpoint returns the same envelope —
      `{ "data": [...], "pagination": { "page": 0, "perPage": 50, "total": 123 } }`.
      `page` is 0-indexed. Pass `?page=` and `?perPage=` as query strings;
      omitting them uses the endpoint's own default page size.
    - **Errors**: non-2xx responses are `{ "error": "human-readable message" }`.
    - **Timestamps**: always ISO-8601 UTC strings (e.g. `2026-08-12T00:32:29.334Z`).
    - **Big numbers as strings**: coin amounts, block heights used as ids,
      registration/deposit/withdrawal ids, and anything else that can exceed
      JavaScript's safe integer range is serialized as a **string**, not a
...
  contact:
    name: SVM Scan
...
servers:
...
security: []
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List transactions
      description: Newest first. Default page size 50.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
        - name: type
          in: query
          description: Filter to transactions containing a message of this exact typeUrl (e.g. `/steemvm.oracle.data.v1.MsgAggregateExchangeRateVote`). See `GET /transactions/message-types` for the full set currently indexed.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - hash: 5CD1859FDC00DBB168787B42F47F2D39658D3F3812E7B831BE9E30933638CA03
                    height: 379199
                    timestamp: '2026-08-13T23:22:00.430Z'
                    code: 0
                    success: true
                    messageTypes: ['/steemvm.steembridge.v1.MsgAttestDeposit']
                    fee: []
                    gasUsed: '68580'
                    gasWanted: '600000'
                pagination: { page: 0, perPage: 1, total: 9064 }
...
  /transactions/{hash}:
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Get a transaction by hash
      description: |
        Messages are decoded server-side (see `@dexplorer/shared`'s message
        registry) — a type this API doesn't have a named decoder for still
        decodes, just with generic `field_1`/`field_2`/... keys instead of
        real names. Five message types get extra enrichment merged into
        `data`: `MsgAttestDeposit`/`MsgSubmitNameRegistration`/
        `MsgAttestWithdrawalPayout`/`MsgAggregateExchangeRateVote`/
        `MsgAggregateExchangeRatePrevote` get an `oracleMoniker` (the
        attesting validator's moniker), and `MsgConfirmName` gets a
        `linkedSteemAccount`. `moduleAccounts` resolves any bech32 address in
        `events` to its Cosmos SDK module account name, where applicable.


        `{hash}` also accepts an EVM-format tx hash (`0x` + 64 hex chars —
        what MetaMask/an EVM wallet links to) for a `MsgEthereumTx`: it's
        resolved to the SVM tx that wraps it via the decoded inner hash, so
        the response's own `hash` field is still the outer Tendermint hash,
        not what was requested. If no such wrapper is indexed yet, this
        endpoint 404s even though the tx may still be visible via
        `GET /evm/tx/{hash}` directly (that endpoint queries the chain's EVM
        JSON-RPC live, independent of this indexer).
      parameters:
        - $ref: '#/components/parameters/hash'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetailResponse'
              example:
                hash: C05E4276B2B53C907E600335B67C3AD7F6A5BA028B87F841D7D65AB542DB5A0F
                height: 353345
                timestamp: '2026-08-12T00:32:29.334Z'
                code: 0
                success: true
                messageTypes: ['/steemvm.steembridge.v1.MsgAttestDeposit']
                fee: []
                gasUsed: '92488'
                gasWanted: '600000'
                log: ''
                memo: ''
                chainId: steemvm
                messages:
                  - typeUrl: /steemvm.steembridge.v1.MsgAttestDeposit
                    data:
                      validator: steem1u96tympny5r8m76psg5jr7fm8kcrg76w3rmpng
                      txid: 5da85737d515500e6cb23a02fd981263d8c7de7f
                      opIndex: 0
                      steemBlock: '108611220'
                      steemTimestamp: '2026-08-12T00:31:36'
                      steemSender: blaze.apps
                      gatewayAccount: svm.bank
                      amountMillisteem: '1000'
                      memo: svm-deposit 0x7fb5a561b74e7da7cd5fb1497830e96b3310d857
                      oracleMoniker: dhaka.witness
                events:
                  - type: deposit_created
                    attributes:
                      - { key: deposit_id, value: '1855' }
                      - { key: txid, value: 5da85737d515500e6cb23a02fd981263d8c7de7f }
                senders: ['steem1u96tympny5r8m76psg5jr7fm8kcrg76w3rmpng']
                ibcTransfer: null
                moduleAccounts:
                  steem10a2y3nc3kqn7cm2lhlkgfgt7du70rstwm0ul3x: steemblackhole
                  steem189zqgygvhf5j56st8698qa9nz6jjhu45d089kz: bridge_reward
                  steem17xpfvakm2amg962yls6f84z3kell8c5lxvg9x8: fee_collector
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    page:
      name: page
      in: query
      description: 0-indexed page number.
      schema: { type: integer, default: 0, minimum: 0 }
    perPage:
      name: perPage
      in: query
      description: Items per page (endpoint-specific default and max).
      schema: { type: integer, minimum: 1, maximum: 200 }
...
    hash:
      name: hash
      in: path
      required: true
      description: Cosmos transaction hash, case-insensitive.
      schema: { type: string }
      example: C05E4276B2B53C907E600335B67C3AD7F6A5BA028B87F841D7D65AB542DB5A0F
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: 'Not found' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        perPage: { type: integer }
        total: { type: integer }
      required: [page, perPage, total]
    Coin:
      type: object
      properties:
        denom: { type: string }
        amount: { type: string }
      required: [denom, amount]
...
    TransactionSummary:
      type: object
      properties:
        hash: { type: string }
        height: { type: integer }
        timestamp: { type: string, format: date-time }
        code: { type: integer, description: 0 means success. }
        success: { type: boolean }
        messageTypes:
          type: array
          items: { type: string }
        fee:
          type: array
          items: { $ref: '#/components/schemas/Coin' }
        gasUsed: { type: string }
        gasWanted: { type: string }
      required: [hash, height, timestamp, code, success, messageTypes, fee, gasUsed, gasWanted]
    DecodeMsg:
      type: object
      description: A decoded Cosmos SDK message. `data`'s shape depends on `typeUrl` — see the message-type registry in `packages/shared/src/encoding/msg.ts`.
      properties:
        typeUrl: { type: string }
        data:
          type: [object, 'null']
          additionalProperties: true
      required: [typeUrl, data]
    EventAttribute:
      type: object
      properties:
        key: { type: string }
        value: { type: string }
        index: { type: boolean }
      required: [key, value]
    TxEvent:
      type: object
      properties:
        type: { type: string }
        attributes:
          type: array
          items: { $ref: '#/components/schemas/EventAttribute' }
      required: [type, attributes]
    IBCTransferFields:
      type: object
      properties:
        sender: { type: string }
        receiver: { type: string }
        sourceChannel: { type: string }
        sourcePort: { type: string }
        destinationChannel: { type: string }
        destinationPort: { type: string }
        amount: { type: string }
        denom: { type: string }
        memo: { type: string }
      required: [sender, receiver, sourceChannel, sourcePort, destinationChannel, destinationPort, amount, denom, memo]
    TransactionDetailResponse:
      allOf:
        - $ref: '#/components/schemas/TransactionSummary'
        - type: object
          properties:
            log: { type: string }
            memo: { type: string }
            chainId: { type: string }
            messages:
              type: array
              items: { $ref: '#/components/schemas/DecodeMsg' }
            events:
              type: array
              items: { $ref: '#/components/schemas/TxEvent' }
            senders:
              type: array
              items: { type: string }
            ibcTransfer:
              oneOf:
                - $ref: '#/components/schemas/IBCTransferFields'
                - type: 'null'
            moduleAccounts:
              type: object
              additionalProperties: { type: string }
              description: bech32 address -> Cosmos SDK module account name (e.g. "fee_collector", "bridge_reward", "steemblackhole"), for labeling module accounts that show up as a sender/receiver in `events`. Empty when STEEMBRIDGE_LCD_URL isn't configured.
          required: [log, memo, chainId, messages, events, senders, ibcTransfer, moduleAccounts]
...